+++ CHANGES.rst
@@ -20,6 +20,7 @@ Changelog
 - added `--log_execution_time` flag to `sub-process`, `tee`, `trigger`
 - `parse_conversion_args` and `perform_pipeline_execution` now support default placeholders HOME/CWD/TMP
 - generators `csv-file`, `dirs`, `files` and `text-file` now support placeholders in their paths
+- added `log-placeholder` filter, which can be used for logging placeholder values
 
 
 0.0.1 (2025-10-31)